refactor: Github.js를 github.js로 변경하여 파일명 대소문자 통일
